Move util function
