+++ packages/web/src/components/track/desktop/ConnectedTrackTile.tsx
@@ -51,8 +51,7 @@ import { isDarkMode, isMatrix } from 'utils/theme/theme'
 
 import {
   getTrackWithFallback,
-  getUserWithFallback,
-  isDescendantElementOf
+  getUserWithFallback
 } from '../helpers'
 import { TrackTileSize } from '../types'
 
@@ -60,6 +59,7 @@ import styles from './ConnectedTrackTile.module.css'
 import TrackTile from './TrackTile'
 import Stats from './stats/Stats'
 import { Flavor } from './stats/StatsText'
+import { isDescendantElementOf } from 'utils/domUtils'
 
 type OwnProps = {
   uid: UID
