+++ packages/web/src/utils/domUtils.ts
@@ -1,3 +1,5 @@
+import { Nullable } from "@audius/common"
+
 export const findAncestor = (el: Element, selector: string) => {
   if (el.closest) {
     return el.closest(selector)
@@ -14,4 +16,13 @@ export const findAncestor = (el: Element, selector: string) => {
     el = el.parentElement
   }
   return el
+}
+
+export const isDescendantElementOf = (
+  descendant: any,
+  ancestor: Nullable<HTMLElement>
+) => {
+  const descendantElement =
+    descendant instanceof Element ? (descendant as Element) : null
+  return ancestor?.contains(descendantElement)
 }
